Adding forgit to the "Version Control "category
Hi,

I am one of the maintainers of forgit. It's a super handy git wrapper that provides interactive ways to add, checkout, stash, and all sorts of other git actions. It uses FZF for fuzzy match, so it makes doing these actions a breeze!

Check it out here, and see if you think it's worthy of being "awesome"!

https://github.com/wfxr/forgit

Thanks :)